Move creation of CheckoutController to separate class
COSDK-1115
