+++ core/src/main/java/com/adyen/checkout/core/components/CheckoutController.kt
@@ -10,29 +10,22 @@ package com.adyen.checkout.core.components
 
 import android.annotation.SuppressLint
 import android.content.Context
-import androidx.appcompat.app.AppCompatDelegate
 import androidx.lifecycle.SavedStateHandle
 import com.adyen.checkout.core.action.data.Action
 import com.adyen.checkout.core.action.internal.ActionComponent
 import com.adyen.checkout.core.action.internal.ActionComponentEvent
 import com.adyen.checkout.core.action.internal.ActionComponentProvider
 import com.adyen.checkout.core.analytics.internal.AnalyticsManager
-import com.adyen.checkout.core.analytics.internal.AnalyticsManagerFactory
-import com.adyen.checkout.core.analytics.internal.AnalyticsSource
 import com.adyen.checkout.core.common.CheckoutContext
 import com.adyen.checkout.core.components.data.model.paymentmethod.PaymentMethods
 import com.adyen.checkout.core.components.internal.PaymentComponentEvent
 import com.adyen.checkout.core.components.internal.PaymentMethodProvider
 import com.adyen.checkout.core.components.internal.ui.PaymentComponent
-import com.adyen.checkout.core.components.internal.ui.model.CommonComponentParamsMapper
 import com.adyen.checkout.core.components.internal.ui.model.ComponentParamsBundle
 import com.adyen.checkout.core.error.toCheckoutError
-import com.adyen.checkout.core.sessions.internal.model.SessionParams
-import com.adyen.checkout.core.sessions.internal.model.SessionParamsFactory
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
-import java.util.Locale
 
 fun CheckoutController(
     target: CheckoutTarget,
@@ -42,55 +35,12 @@ fun CheckoutController(
     applicationContext: Context,
     coroutineScope: CoroutineScope,
 ): CheckoutController {
-    val checkoutConfiguration: CheckoutConfiguration
-    val checkoutAttemptId: String?
-    val publicKey: String?
-    val componentSessionParams: SessionParams?
-    val sessionId: String?
-
-    when (context) {
-        is CheckoutContext.Advanced -> {
-            checkoutConfiguration = context.checkoutConfiguration
-            checkoutAttemptId = context.checkoutAttemptId
-            publicKey = context.publicKey
-            componentSessionParams = null
-            sessionId = null
-        }
-
-        is CheckoutContext.Sessions -> {
-            checkoutConfiguration = context.checkoutConfiguration
-            checkoutAttemptId = context.checkoutAttemptId
-            publicKey = context.publicKey
-            componentSessionParams = SessionParamsFactory.create(context.checkoutSession)
-            sessionId = context.checkoutSession.sessionSetupResponse.id
-        }
-    }
-
-    val componentParamsBundle = CommonComponentParamsMapper().mapToParams(
-        checkoutConfiguration = checkoutConfiguration,
-        deviceLocale = AppCompatDelegate.getApplicationLocales()[0] ?: Locale.getDefault(),
-        dropInOverrideParams = null,
-        componentSessionParams = componentSessionParams,
-        publicKey = publicKey,
-    )
-
-    val analyticsManager = AnalyticsManagerFactory().provide(
-        componentParams = componentParamsBundle.commonComponentParams,
-        applicationContext = applicationContext,
-        // TODO - Analytics: Pass the correct paymentMethod type
-        source = AnalyticsSource.PaymentComponent("paymentMethod.type"),
-        sessionId = sessionId,
-        checkoutAttemptId = checkoutAttemptId,
-    )
-
-    return CheckoutController(
+    return CheckoutControllerFactory().create(
         target = target,
         context = context,
         callbacks = callbacks,
+        applicationContext = applicationContext,
         coroutineScope = coroutineScope,
-        analyticsManager = analyticsManager,
-        checkoutConfiguration = checkoutConfiguration,
-        componentParamsBundle = componentParamsBundle,
     )
 }
